# fix/test: DistanceTransform array out of bounds exception