* adds tests that reproduce the error in #108