@bogovicj 2025-03-18
The second commit to this PR substantially speeds things up (40% - 100%), but using primitive arrays for parts of the  lowest level methods `transformAlongDimension` and `transformAlongDimensionPropagateLabels`

Because this implementation makes an attempt to support very large images, we fall back to the old implementation if any dimension is longer than `Integer.MAX_VALUE`. 

That commit also includes a benchmark, some of the results of which are shown below:


# Selected benchmark results
 
## 2D

### old single thread
```
Benchmark                                            Mode  Cnt    Score   Error  Units
DistanceTransformBenchmark.voronoiDistanceTransform  avgt   25  484.629 ± 6.081  ms/op
```
### new single thread
```
Benchmark                                            Mode  Cnt    Score   Error  Units
DistanceTransformBenchmark.voronoiDistanceTransform  avgt   25  295.930 ± 3.423  ms/op
```

### old multi-thread (8)
```
Benchmark                                            Mode  Cnt    Score   Error  Units
DistanceTransformBenchmark.voronoiDistanceTransform  avgt   25  170.793 ± 1.800  ms/op
```
### new multi-thread (8)
```
Benchmark                                            Mode  Cnt    Score   Error  Units
DistanceTransformBenchmark.voronoiDistanceTransform  avgt   25  118.340 ± 2.338  ms/op
```

## 3D

### old single threaded

```
Benchmark                                            Mode  Cnt     Score    Error  Units
DistanceTransformBenchmark.distanceTransform         avgt   25  1167.336 ± 14.451  ms/op
DistanceTransformBenchmark.voronoiDistanceTransform  avgt   25  1852.702 ± 20.670  ms/op
```

### new single threaded
```
Benchmark                                            Mode  Cnt     Score    Error  Units
DistanceTransformBenchmark.distanceTransform         avgt   25   605.162 ±  6.575  ms/op
DistanceTransformBenchmark.voronoiDistanceTransform  avgt   25  1315.934 ± 22.510  ms/op
```